+++ smart-contracts/wasm-chain-integration/CHANGELOG.md
@@ -2,9 +2,15 @@
 
 ## Unreleased changes
 
+## concordium-smart-contract-engine 1.1.0 (2023-04-12)
+
+- Add `saturating_sub` method to `InterpreterEnergy`.
+- Derive `PartialEq`/`Eq` for `Logs` and `InvokeFailure`.
+
 ## concordium-smart-contract-engine 1.0.1 (2023-03-20)
 
 - Bump concordium-contracts-common dependency to 5.3.
+
 
 ## concordium-smart-contract-engine 1.0.0 (2023-02-03)